gfx/layers/apz/test/mochitest/test_group_keyboard.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Various keyboard scrolling tests</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="apz_test_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

var subtests = [
  {"file": "helper_key_scroll.html", prefs: [["apz.test.logging_enabled", true],
                                             ["test.events.async.enabled", true]]},
  {"file": "helper_bug1674935.html", prefs: []},
  {"file": "helper_relative_scroll_smoothness.html",
   prefs: [
    ["apz.test.logging_enabled", true],
    // Use a longer animation duration to avoid the situation that the
    // animation stops accidentally in between each arrow down key press.
    // If the situation happens, scroll offsets will not change at the moment.
    ["general.smoothScroll.lines.durationMaxMS", 1500],
    ["general.smoothScroll.lines.durationMinMS", 1500]
   ]
  },
];

    if (isKeyApzEnabled()) {
      SimpleTest.waitForExplicitFinish();
      window.onload = function() {
        runSubtestsSeriallyInFreshWindows(subtests)
        .then(SimpleTest.finish, SimpleTest.finishWithFailure);
      };
    } else {
      SimpleTest.ok(true, "Keyboard APZ is disabled");
    }
  </script>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1383365">Async key scrolling test</a>
</body>
</html>
